re-organize content pages
