+++ kolibri/plugins/style_guide/assets/src/views/content/checkboxes/index.vue
