+++ kolibri/plugins/style_guide/assets/src/views/content/modals/index.vue
