+++ kolibri/plugins/style_guide/assets/src/views/content/navbar/example.html
