+++ kolibri/plugins/style_guide/assets/src/views/content/style-guide-home/index.vue
