+++ kolibri/plugins/style_guide/assets/src/views/shell/nav-menu.js
@@ -1,15 +1,15 @@
 import { sortBy, flatten } from 'lodash';
 
-import kButton from '../content/components/k-button';
-import typography from '../content/style/typography';
-import checkboxes from '../content/components/checkboxes';
-import radioButtons from '../content/components/radio-buttons';
-import textFields from '../content/components/text-fields';
-import simpleModals from '../content/components/simple-modals';
-import filters from '../content/components/filters';
-import kBreadcrumbs from '../content/components/k-breadcrumbs';
-import kNavbar from '../content/components/k-navbar';
-import homePage from '../content/home/home-page';
+import buttonsPage from '../content/buttons';
+import typographyPage from '../content/typography';
+import checkboxesPage from '../content/checkboxes';
+import radioButtonsPage from '../content/radio-buttons';
+import textFieldsPage from '../content/text-fields';
+import modalsPage from '../content/modals';
+import filtersPage from '../content/filters';
+import breadcrumbsPage from '../content/breadcrumbs';
+import navbarPage from '../content/navbar';
+import homePage from '../content/style-guide-home';
 
 function sortSectionItems(items) {
   return sortBy(items, [item => item.itemName]);
@@ -27,20 +27,6 @@ const homeRoute = [
 // Notes: This is view-agnostic; it doesn't make assumption on how it will be
 // rendered (whether it's a side-nav or a horizontal menu).
 const navMenu = [
-  /*
-  {
-    sectionName: 'Styles',
-    sectionItems: sortSectionItems([
-      {
-        itemName: 'Typography',
-        itemRoute: {
-          path: `${BASE}/typography`,
-          component: typography,
-        },
-      },
-    ]),
-  },
-  */
   {
     sectionName: 'Components',
     sectionItems: sortSectionItems([
@@ -48,60 +34,23 @@ const navMenu = [
         itemName: 'Buttons',
         itemRoute: {
           path: `/components/buttons`,
-          component: kButton,
+          component: buttonsPage,
         },
       },
       {
         itemName: 'Breadcrumbs',
         itemRoute: {
           path: `/components/breadcrumbs`,
-          component: kBreadcrumbs,
+          component: breadcrumbsPage,
         },
       },
       {
         itemName: 'Horizontal Navbar',
         itemRoute: {
           path: `/components/navbar`,
-          component: kNavbar,
+          component: navbarPage,
         },
       },
-      /*
-      {
-        itemName: 'Checkboxes',
-        itemRoute: {
-          path: `${BASE}/components/checkboxes`,
-          component: checkboxes,
-        },
-      },
-      {
-        itemName: 'Radio buttons',
-        itemRoute: {
-          path: `${BASE}/components/radio-buttons`,
-          component: radioButtons,
-        },
-      },
-      {
-        itemName: 'Text fields',
-        itemRoute: {
-          path: `${BASE}/components/text-fields`,
-          component: textFields,
-        },
-      },
-      {
-        itemName: 'Simple modals',
-        itemRoute: {
-          path: `${BASE}/components/simple-modals`,
-          component: simpleModals,
-        },
-      },
-      {
-        itemName: 'Filters',
-        itemRoute: {
-          path: `${BASE}/components/filters`,
-          component: filters,
-        },
-      },
-      */
     ]),
   },
 ];
